Ad Updated Issue, IssueList and Deleted SearchForm Because: - Moved the codes from SearchForm to IssueList to have all of the useEffects and make the search functionality work.
